enable prettier

+++ src/chains/ethereum/__tests__/api/evm/snapshot.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.7.4;
 
 contract snapshot {
-    uint public n = 42;
+    uint256 public n = 42;
 
     function inc() public {
         n += 1;
