+++ src/chains/ethereum/__tests__/contracts/HelloWorld.sol
@@ -2,20 +2,20 @@
 pragma solidity ^0.7.4;
 
 contract HelloWorld {
-  uint public value;
+    uint256 public value;
 
-  event ValueSet(uint);
+    event ValueSet(uint256);
 
-  constructor() payable {
-    value = 5;
-  }
+    constructor() payable {
+        value = 5;
+    }
 
-  function setValue(uint val) public {
-    value = val;
-    emit ValueSet(val);
-  }
+    function setValue(uint256 val) public {
+        value = val;
+        emit ValueSet(val);
+    }
 
-  function getConstVal() public pure returns (uint8) {
-    return 123;
-  }
+    function getConstVal() public pure returns (uint8) {
+        return 123;
+    }
 }
